requirements.txt
aiodns==1.1.1
aiohttp==3.0.1
async-timeout==2.0.1
attrs==17.4.0
cchardet==2.1.1
ccxt==1.11.149
certifi==2018.1.18
chardet==3.0.4
cryptg==0.1.0.7
cryptocompare==0.6
cycler==0.10.0
emoji==0.4.5
idna==2.6
idna-ssl==1.0.1
kiwisolver==1.0.1
matplotlib==2.2.2
multidict==4.1.0
numpy==1.14.2
pandas==0.22.0
pyaes==1.6.1
pyasn1==0.4.2
pycares==2.3.0
pyparsing==2.2.0
python-dateutil==2.7.1
python-dotenv==0.8.2
pytz==2018.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
Telethon==0.18.1.1
urllib3==1.22
yarl==1.1.0
